Add Unmployment.css styling
